# contracts/openapi/public.yaml
openapi: 3.1.0
info:
  title: Skill Hub Public API
  version: 0.1.0
  description: |
    Web to Go control plane API. This file is the single source of truth for the
    contract (implementation rule 12); the TS client under packages/ is generated
    from it and must never be hand-edited.
servers:
  - url: http://localhost:8080

# Every route in internal/apiserver/router.go is wrapped in RequireSession
# unless it is listed below with its own `security`. Inheriting the requirement
# here rather than repeating it per operation keeps the two in step: a new
# operation is authenticated by default, which is the direction a mistake should
# fall in.
security:
  - sessionCookie: []
paths:
...
  /auth/github/callback:
    get:
      operationId: finishGithubLogin
      summary: GitHub OAuth callback; creates user and workspace on first login
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "302":
          description: Login succeeded; session cookie set, redirect to the app.
        "401":
          description: State mismatch or code exchange failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/dev/login:
    post:
      operationId: devLogin
      summary: Offline dev-provider login (ADR-020)
      description: |
        Exists only when the server runs with DEV_LOGIN=1 (local demo and E2E);
        in production the route is absent and returns 404.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  maxLength: 64
                  description: Dev user handle; defaults to "dev".
      responses:
        "204":
          description: Logged in; session cookie set.
...
  /me:
    get:
      operationId: getMe
      summary: Current user and personal workspace, resolved from the session
      responses:
        "200":
          description: The authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: requestAccountDeletion
      summary: Request deletion of the caller's account (CORE-007)
      description: |
        Starts a 30-day grace period; nothing is removed yet and the account
        stays usable so the request can be cancelled. Idempotent — asking twice
        keeps the original start time rather than extending the wait.

        When the grace period ends, uploaded datasets, run and download
        artifacts, and every skill nobody else forked or ran are permanently
        deleted, files included. Skill versions other users forked or historical
        runs used are retained with the owner's identity removed (PDM-006 §6.1):
        deleting them would break a third party's provenance chain, which
        DISC-003 requires to stay intact.
      responses:
        "200":
          description: Grace period started, with the scope of what will be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletion"
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/deletion/cancel:
    post:
      operationId: cancelAccountDeletion
      summary: Withdraw a pending account deletion (CORE-007)
      description: |
        Valid for the whole grace period. Idempotent: cancelling when nothing is
        pending is a no-op.
      responses:
        "200":
          description: No deletion is pending any more.
          content:
            application/json:
              schema:
                type: object
                required: [deletion_requested_at]
                properties:
                  deletion_requested_at:
                    type: [string, "null"]
                    format: date-time
                    description: |
                      Always null here - the field exists so a client can read
                      the post-cancel state from the same shape it reads
                      elsewhere, rather than inferring it from a bare 200.
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sh_session
      description: |
        The opaque session token set by the auth routes and resolved by
        RequireSession (apps/platform/internal/identity/http.go). HttpOnly,
        SameSite=Lax, Secure outside plain-http local dev, so a browser sends it
        on its own and no client code reads or sets it - the generated TS client
        only has to send credentials with the request.

        It identifies the user and nothing more. The workspace scope every query
        runs under is resolved from it server-side and is never taken from a
        request parameter (iron rule 3, ADR-011).
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Me:
      type: object
      required: [user_id, email, display_name, workspace_id, deletion_requested_at, purge_after]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        workspace_id:
          type: string
          format: uuid
        deletion_requested_at:
          type: [string, "null"]
          format: date-time
          description: |
            When the caller asked for their account to be deleted, or null when
            no deletion is pending. 02:SEC-006 requires the deletion job to have
            a state the user can follow, and until this field existed the only
            place it appeared was the response to DELETE /me itself — a user who
            closed the tab could not ask again.
        purge_after:
          type: [string, "null"]
          format: date-time
          description: |
            When the grace period ends and the purge becomes due; cancellable
            until then. Null exactly when `deletion_requested_at` is null.
    AccountDeletion:
      type: object
      required: [deletion_requested_at, purge_after, cancellable, scope]
      properties:
        deletion_requested_at:
          type: string
          format: date-time
        purge_after:
          type: string
          format: date-time
          description: When the purge becomes due; cancellable until then.
        cancellable:
          type: boolean
        scope:
          type: string
          description: |
            Plain-language statement of what is deleted and what is retained
            de-identified. WS-002 requires the scope to be shown before the
            deletion, not discovered after it.
...